Move `CheckMachineDeployment` function to `health` package
